+++ app/mailers/notifier.rb
@@ -55,53 +55,19 @@ class Notifier < ActionMailer::Base
     end
   end
 
-  def started_sharing(recipient_id, sender_id)
-    send_notification(:started_sharing, recipient_id, sender_id)
-  end
-
-  def liked(recipient_id, sender_id, like_id)
-    send_notification(:liked, recipient_id, sender_id, like_id)
-  end
-
-  def reshared(recipient_id, sender_id, reshare_id)
-    send_notification(:reshared, recipient_id, sender_id, reshare_id)
-  end
-
-  def mentioned(recipient_id, sender_id, target_id)
-    send_notification(:mentioned, recipient_id, sender_id, target_id)
-  end
-
-  def comment_on_post(recipient_id, sender_id, comment_id)
-    send_notification(:comment_on_post, recipient_id, sender_id, comment_id)
-  end
-
-  def also_commented(recipient_id, sender_id, comment_id)
-    send_notification(:also_commented, recipient_id, sender_id, comment_id)
-  end
-
-  def private_message(recipient_id, sender_id, message_id)
-    send_notification(:private_message, recipient_id, sender_id, message_id)
-  end
-
-  def confirm_email(recipient_id)
-    send_notification(:confirm_email, recipient_id)
-  end
-
-  def csrf_token_fail(recipient_id)
-    send_notification(:csrf_token_fail, recipient_id)
-  end
-
-  private
   def send_notification(type, *args)
     @notification = NotificationMailers.const_get(type.to_s.camelize).new(*args)
 
     with_recipient_locale do
       mail(@notification.headers) do |format|
+        self.action_name = type
         format.text
         format.html
       end
     end
   end
+
+  private
 
   def with_recipient_locale(&block)
     I18n.with_locale(@notification.recipient.language, &block)