+++ app/workers/mail/csrf_token_fail.rb
@@ -1,11 +1,6 @@
 module Workers
   module Mail
-    class CsrfTokenFail < Base
-      sidekiq_options queue: :low
-
-      def perform(user_id)
-        Notifier.csrf_token_fail(user_id).deliver_now
-      end
+    class CsrfTokenFail < NotifierBase
     end
   end
 end