+++ spec/workers/mail/mentioned_spec.rb
@@ -3,15 +3,16 @@
 #   the COPYRIGHT file.
 
 describe Workers::Mail::Mentioned do
-  describe '#perfom' do
-    it 'should call .deliver on the notifier object' do
+  describe "#perform" do
+    it "should call .deliver on the notifier object" do
       user = alice
       sm = FactoryGirl.build(:status_message)
-      m = Mention.new(:person => user.person, :post=> sm)
+      m = Mention.new(person: user.person, mentions_container: sm)
 
       mail_double = double()
       expect(mail_double).to receive(:deliver_now)
-      expect(Notifier).to receive(:mentioned).with(user.id, sm.author.id, m.id).and_return(mail_double)
+      expect(Notifier).to receive(:send_notification)
+        .with("mentioned", user.id, sm.author.id, m.id).and_return(mail_double)
 
       Workers::Mail::Mentioned.new.perform(user.id, sm.author.id, m.id)
     end