+++ Changelog.md
@@ -5,6 +5,7 @@
 ## Bug fixes
 
 ## Features
+* Add support for mentions in comments to the backend [#6818](https://github.com/diaspora/diaspora/pull/6818)
 
 # 0.6.2.0